# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Stock Manager API",
        "description": "API documentation for the Laravel Stock Manager application.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/entries": {
            "get": {
                "tags": [
                    "Entries"
                ],
                "summary": "Get all entries",
                "operationId": "92f5d5ab7e3400ceb5bf9a813e51af92",
                "responses": {
                    "200": {
                        "description": "List of all entries",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Entry"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Entries"
                ],
                "summary": "Create a new entry",
                "operationId": "0b8151ed9b9d29c014da7939bf778e9a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateEntryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Entry created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Entry"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/entries/{id}": {
            "get": {
                "tags": [
                    "Entries"
                ],
                "summary": "Get an entry by ID",
                "operationId": "8d5db05ac5b117e1b7730f6f44950cb8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the entry",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Entry details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Entry"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/entries/product/{productId}": {
            "get": {
                "tags": [
                    "Entries"
                ],
                "summary": "Get entries by product ID",
                "operationId": "24bee7d6469afcb7714492691ca4b87d",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "ID of the product",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of entries for the specified product",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Entry"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/entries/date-range": {
            "post": {
                "tags": [
                    "Entries"
                ],
                "summary": "Get entries by date range",
                "operationId": "84829eb45e7d7b94d9e9733a8b5927f3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "start_date",
                                    "end_date"
                                ],
                                "properties": {
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-01"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-31"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "List of entries within the specified date range",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Entry"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/outputs": {
            "get": {
                "tags": [
                    "Outputs"
                ],
                "summary": "Get all outputs",
                "operationId": "05779a1e48e7e4974ebba5d33120c2dc",
                "responses": {
                    "200": {
                        "description": "List of all outputs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Output"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Outputs"
                ],
                "summary": "Create a new output",
                "operationId": "e979ad6387804d0991ae96f08b7f9321",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOutputRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Output created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Output"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/outputs/{id}": {
            "get": {
                "tags": [
                    "Outputs"
                ],
                "summary": "Get an output by ID",
                "operationId": "5bc91ca1ac3a4f9c6d3f80895f56c8cf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the output",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Output details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Output"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/outputs/product/{productId}": {
            "get": {
                "tags": [
                    "Outputs"
                ],
                "summary": "Get outputs by product ID",
                "operationId": "c711b0409bb83b5f9d596a5e5bd7134b",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "ID of the product",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of outputs for the specified product",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Output"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/outputs/date-range": {
            "post": {
                "tags": [
                    "Outputs"
                ],
                "summary": "Get outputs by date range",
                "operationId": "514696888cd6c4b7df523e96baa66c5c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "start_date",
                                    "end_date"
                                ],
                                "properties": {
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-01"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-31"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "List of outputs within the specified date range",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Output"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateEntryRequest": {
                "required": [
                    "product_id",
                    "quantity"
                ],
                "properties": {
                    "product_id": {
                        "type": "string",
                        "example": "1"
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 50
                    }
                },
                "type": "object"
            },
            "CreateOutputRequest": {
                "required": [
                    "product_id",
                    "quantity"
                ],
                "properties": {
                    "product_id": {
                        "type": "string",
                        "example": "1"
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 20
                    }
                },
                "type": "object"
            },
...
            "Entry": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "1"
                    },
                    "product_id": {
                        "type": "string",
                        "example": "1"
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 50
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-02T12:00:00Z"
                    }
                },
                "type": "object"
            },
            "Output": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "1"
                    },
                    "product_id": {
                        "type": "string",
                        "example": "1"
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 20
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-02T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter token in format (Bearer <token>)"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}